Test both utp and tcp with another client
