+++ crates/librqbit/src/tests/mod.rs
@@ -1,4 +1,4 @@
 mod e2e;
+mod e2e_another_local_client;
 mod e2e_stream;
-mod e2e_utp;
 pub mod test_util;
